pyproject.toml
[project]
name = "fast-voice"
version = "0.1.0"
description = "Fast, split-architecture voice-to-text system using faster-whisper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=14.2",
    "pyaudio>=0.2.14",
    "numpy>=2.2.2",
    "pyyaml>=6.0.2",
    "faster-whisper>=1.1.1",
    "torch>=2.6.0",
    "huggingface-hub>=0.28.1",
    # whisper-live dependencies (scipy is enough for most backends)
    "scipy>=1.11.0"
]

[project.scripts]
fast-voice-server = "fast_voice.server.main:start"
fast-voice-client = "fast_voice.client.tui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fast_voice"]
